test: move session store pruning integration suite to e2e lane
